tests done, starting with project now
